+++ app/src/main/java/com/example/treasurehunt/MainActivity.kt
@@ -25,14 +25,14 @@ import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.treasurehunt.data.ScreenList
-import com.example.treasurehunt.screens.AchievementsScreen
-import com.example.treasurehunt.screens.EndGameScreen
-import com.example.treasurehunt.screens.HomeScreen
-import com.example.treasurehunt.screens.OnboardingScreen
-import com.example.treasurehunt.screens.PlayGameScreen
-import com.example.treasurehunt.screens.RuleScreen
-import com.example.treasurehunt.screens.StartGameScreen
+import com.example.treasurehunt.ui.ScreenList
+import com.example.treasurehunt.ui.screens.AchievementsScreen
+import com.example.treasurehunt.ui.screens.EndGameScreen
+import com.example.treasurehunt.ui.screens.HomeScreen
+import com.example.treasurehunt.ui.screens.OnboardingScreen
+import com.example.treasurehunt.ui.screens.PlayGameScreen
+import com.example.treasurehunt.ui.screens.RuleScreen
+import com.example.treasurehunt.ui.screens.StartGameScreen
 import com.example.treasurehunt.ui.theme.TreasureHuntTheme
 import com.example.treasurehunt.utils.AppUtils
 import com.example.treasurehunt.utils.Response
