+++ app/src/main/java/com/example/treasurehunt/model/CompletedQuestData.kt
@@ -1,0 +1,8 @@
+package com.example.treasurehunt.model
+
+import androidx.annotation.StringRes
+
+data class CompletedQuestData(
+    @StringRes val questDetail: Int,
+    val questPicture: QuestImage
+)
