+++ app/src/main/java/com/example/treasurehunt/model/QuestItem.kt
@@ -1,0 +1,11 @@
+package com.example.treasurehunt.model
+
+data class QuestItem(
+    val title: String,
+    val description: String,
+    val clue: String,
+    val hint: String,
+    val coordinates: Coordinate,
+    val rating: Double,
+    val endGameAssets: CompletedQuestData
+)
