+++ app/src/main/java/com/example/treasurehunt/ui/screens/PermissionAlertDialog.kt
@@ -1,4 +1,4 @@
-package com.example.treasurehunt.screens
+package com.example.treasurehunt.ui.screens
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Text
