+++ app/src/main/java/com/example/treasurehunt/ui/state/StartGameState.kt
@@ -1,0 +1,15 @@
+package com.example.treasurehunt.ui.state
+
+import com.example.treasurehunt.GetGreetingQuery
+import com.example.treasurehunt.model.Coordinate
+import com.example.treasurehunt.model.QuestItem
+
+data class StartGameState(
+    val greetings: List<GetGreetingQuery.Greeting?>? = null,
+    val currentLocation: Coordinate? = null,
+    val questRadius: Double? = null,
+    val questList: List<QuestItem>? = null
+)
+
+
+
